[chore] renaming namespaces
