+++ backend/LogSummitApi.Application/Core/Factories/ValidatorFactory.cs
@@ -1,8 +1,8 @@
 ﻿using LogSummitApi.Application.Core.Validators;
 using LogSummitApi.Domain.Core.Entities;
+using LogSummitApi.Domain.Core.Interfaces.Common;
 using LogSummitApi.Domain.Core.Interfaces.Factories;
 using LogSummitApi.Domain.Core.Interfaces.Repositories;
-using LogSummitApi.Domain.Core.Interfaces.Utilities;
 
 namespace LogSummitApi.Application.Core.Factories;
