+++ backend/LogSummitApi.Application/Core/Utilities/Jwt.cs
@@ -1,7 +1,7 @@
 ﻿using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
-using LogSummitApi.Domain.Core.Interfaces.Utilities;
+using LogSummitApi.Domain.Core.Interfaces.Common;
 using Microsoft.IdentityModel.Tokens;
 
 namespace LogSummitApi.Application.Core.Utilities;
