+++ backend/LogSummitApi.Domain/Core/Interfaces/Common/IHasher.cs
@@ -1,4 +1,4 @@
-﻿namespace LogSummitApi.Domain.Core.Interfaces.Utilities;
+﻿namespace LogSummitApi.Domain.Core.Interfaces.Common;
 
 public interface IHasher
 {
