+++ backend/LogSummitApi.Domain/Core/Interfaces/Common/IValidator.cs
@@ -1,4 +1,4 @@
-﻿namespace LogSummitApi.Domain.Core.Interfaces.Utilities;
+﻿namespace LogSummitApi.Domain.Core.Interfaces.Common;
 
 public interface IValidator<T>
 {
