+++ backend/LogSummitApi.Domain/Core/Interfaces/Factories/IValidatorFactory.cs
@@ -1,5 +1,5 @@
 ﻿using LogSummitApi.Domain.Core.Entities;
-using LogSummitApi.Domain.Core.Interfaces.Utilities;
+using LogSummitApi.Domain.Core.Interfaces.Common;
 
 namespace LogSummitApi.Domain.Core.Interfaces.Factories;
